requirements.txt
Django==3.2.5
django_heroku==0.3.1
jsons==1.4.1
gunicorn
python-decouple
uvicorn
uvloop
httptools
django-cors-headers
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
PyJWT==2.1.0
pytz==2021.1
newrelic